+++ src/pvenv/subcommands/base.py
@@ -2,8 +2,9 @@ from argparse import Namespace
 
 
 class BaseCommand:
-    def __init__(self, _options: Namespace):
+    def __init__(self, options: Namespace):
         self._prefix = "_pvenv_env"
+        self._options = options
 
     def run(self):
         raise NotImplementedError(f"{self.__class__.__name__} must implement run")