+++ CHANGELOG.rst
@@ -9,6 +9,12 @@ The format is based on `Keep a Changelog`_, and this project adheres to `Semanti
 `Unreleased`_
 -------------
 
+`0.3.1`_ - 2022-22-17
+---------------------
+Fixed
+^^^^^
+* Removed breaking bug in mkvenv
+
 `0.3.0`_ - 2022-22-17
 ---------------------
 Added
@@ -31,7 +37,8 @@ Added
 ^^^^^
 * lsvenv/venv list to list all the available venvs
 
-.. _`unreleased`: https://github.com/spapanik/pvenv/compare/v0.3.0...main
+.. _`unreleased`: https://github.com/spapanik/pvenv/compare/v0.3.1...main
+.. _`0.3.1`: https://github.com/spapanik/pvenv/compare/v0.3.0...v0.3.1
 .. _`0.3.0`: https://github.com/spapanik/pvenv/compare/v0.2.0...v0.3.0
 .. _`0.2.0`: https://github.com/spapanik/pvenv/compare/v0.1.0...v0.2.0
 .. _`0.1.0`: https://github.com/spapanik/yamk/releases/tag/v0.1.0